requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
email-validator
Flask-Uploads
Pillow
